feat(api): lock_duration --> lock_time, moving new endpoint to /stake/
