+++ api/registry/models.py
@@ -192,76 +192,3 @@ class GTCStakeEvent(models.Model):
                 name="gtc_staking_index_by_staker",
             ),
         ]
-
-
-# Stores the current state of each stake
-class Stake(models.Model):
-    class Chain(models.TextChoices):
-        ETHEREUM = 0, "Ethereum Mainnet"
-        OPTIMISM = 1, "Optimism Mainnet"
-
-    chain = models.SmallIntegerField(
-        choices=Chain.choices, default=Chain.ETHEREUM, db_index=True
-    )
-    unlock_time = models.DateTimeField(null=False, blank=False)
-
-    # u256
-    last_updated_in_block = models.DecimalField(
-        decimal_places=0, null=False, blank=False, max_digits=78, db_index=True
-    )
-
-    # u64
-    lock_duration = models.DecimalField(
-        decimal_places=0, null=False, blank=False, max_digits=20
-    )
-
-    # For self-stake, staker and stakee are the same
-    staker = EthAddressField(null=False, blank=False, db_index=True)
-    stakee = EthAddressField(null=False, blank=False, db_index=True)
-
-    # u256
-    current_amount = models.DecimalField(
-        decimal_places=0, null=False, blank=False, max_digits=78
-    )
-
-    class Meta:
-        unique_together = ["staker", "stakee", "chain"]
-
-
-# Stores raw staking events, for analysis and debugging
-class StakeEvent(models.Model):
-    class StakeEventType(models.TextChoices):
-        SELF_STAKE = "SST"
-        COMMUNITY_STAKE = "CST"
-        SELF_STAKE_WITHDRAW = "SSW"
-        COMMUNITY_STAKE_WITHDRAW = "CSW"
-        SLASH = "SLA"
-        RELEASE = "REL"
-
-    event_type = models.CharField(
-        max_length=3,
-        choices=StakeEventType.choices,
-        blank=False,
-        db_index=True,
-    )
-
-    chain = models.SmallIntegerField(
-        choices=Stake.Chain.choices, null=False, blank=False, db_index=True
-    )
-
-    # For self-stake, staker and stakee are the same
-    staker = EthAddressField(null=False, blank=False, db_index=True)
-    stakee = EthAddressField(null=False, blank=False, db_index=True)
-
-    amount = models.DecimalField(
-        decimal_places=0, null=False, blank=False, max_digits=78
-    )
-
-    block_number = models.DecimalField(
-        decimal_places=0, null=False, blank=False, max_digits=78
-    )
-
-    tx_hash = models.CharField(max_length=66, null=False, blank=False)
-
-    # Only applies to SelfStake and CommunityStake events
-    unlock_time = models.DateTimeField(null=True, blank=True)
