+++ api/stake/test/conftest.py
@@ -1,0 +1,2 @@
+# pylint: disable=unused-import
+from scorer.test.conftest import sample_address, sample_token
