+++ api/stake/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .api import api
+
+urlpatterns = [
+    path("", api.urls),
+]
