+++ indexer/src/staking_indexer.rs
@@ -130,7 +130,7 @@ impl<'a> StakingIndexer<'a> {
             match previous_events_query {
                 Ok(previous_events) => {
                     for (event, meta) in previous_events.iter() {
-                        self.process_staking_event(&event, &meta).await?;
+                        self.process_staking_event(&event, &meta, &client).await?;
                     }
                 }
                 Err(err) => {
@@ -171,7 +171,7 @@ impl<'a> StakingIndexer<'a> {
                 Ok(event_with_meta) => event_with_meta,
             };
 
-            self.process_staking_event(&event, &meta).await?;
+            self.process_staking_event(&event, &meta, &client).await?;
         }
 
         Ok(())
@@ -181,17 +181,18 @@ impl<'a> StakingIndexer<'a> {
         &self,
         event: &IdentityStakingEvents,
         meta: &LogMeta,
+        client: &Arc<ethers::providers::Provider<ethers::providers::Ws>>,
     ) -> Result<()> {
         let block_number = meta.block_number.as_u64();
         let tx_hash = format!("{:?}", meta.transaction_hash);
 
         match event {
             IdentityStakingEvents::SelfStakeFilter(event) => {
-                self.process_self_stake_event(&event, block_number, &tx_hash)
+                self.process_self_stake_event(&event, block_number, &tx_hash, &client)
                     .await
             }
             IdentityStakingEvents::CommunityStakeFilter(event) => {
-                self.process_community_stake_event(&event, block_number, &tx_hash)
+                self.process_community_stake_event(&event, block_number, &tx_hash, &client)
                     .await
             }
             IdentityStakingEvents::SelfStakeWithdrawnFilter(event) => {
@@ -220,12 +221,31 @@ impl<'a> StakingIndexer<'a> {
         }
     }
 
+    // Could cache this somehow to avoid checking timestamp multiple times
+    // if there are multiple stake events in the same block
+    async fn get_timestamp_for_block_number(
+        &self,
+        block_number: u64,
+        client: &Arc<ethers::providers::Provider<ethers::providers::Ws>>,
+    ) -> Result<u64> {
+        if let Ok(Some(block)) = client.get_block(block_number).await {
+            return Ok(block.timestamp.as_u64())
+        }
+        Err(eyre::eyre!(
+            "Failed to fetch block timestamp for block {}",
+            block_number
+        ))
+    }
+
     async fn process_self_stake_event(
         &self,
         event: &SelfStakeFilter,
         block_number: u64,
         tx_hash: &String,
+        client: &Arc<ethers::providers::Provider<ethers::providers::Ws>>,
     ) -> Result<()> {
+        let timestamp = self.get_timestamp_for_block_number(block_number, client).await?;
+
         if let Err(err) = self
             .postgres_client
             .add_or_extend_stake(
@@ -235,7 +255,7 @@ impl<'a> StakingIndexer<'a> {
                 &event.staker,
                 &event.amount,
                 &event.unlock_time,
-                &event.lock_duration,
+                &timestamp,
                 &block_number,
                 tx_hash,
             )
@@ -254,7 +274,10 @@ impl<'a> StakingIndexer<'a> {
         event: &CommunityStakeFilter,
         block_number: u64,
         tx_hash: &String,
+        client: &Arc<ethers::providers::Provider<ethers::providers::Ws>>,
     ) -> Result<()> {
+        let timestamp = self.get_timestamp_for_block_number(block_number, client).await?;
+
         if let Err(err) = self
             .postgres_client
             .add_or_extend_stake(
@@ -264,7 +287,7 @@ impl<'a> StakingIndexer<'a> {
                 &event.stakee,
                 &event.amount,
                 &event.unlock_time,
-                &event.lock_duration,
+                &timestamp,
                 &block_number,
                 tx_hash,
             )
